Renamed the project to 'Jigsaw' and made some preliminary changes, preparing for fancy torrent name rendering.
git-svn-id: file:///var/svn-repos/cocoa/mrtorrent/trunk/mrtorrent@40 bf3e2900-abe3-4bc4-868a-2698488a2087

+++ TorrentModel.m
@@ -95,7 +95,7 @@
 
 - (double) proportionComplete
 {
-	return downloaded / size;
+	return (double)downloaded / (double)size;
 }
 
 - (unsigned long long) secondsRemaining
